fix: split in-flight work query tests for backend-size limit
CI backend-size failed because server_work_query_test.go exceeded the
1000-line file cap after dispatch materialization API tests.
